moving code to paasresource module
