+++ internal/paasresource/main.go
@@ -1,4 +1,4 @@
-package v1alpha1
+package paasresource
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
